chore(stepfunctions-tasks): move stepfunctions start-execution integration into directory
We recently re-organized service-integrations by creating a folder for each service that
stepfunctions integrates with and missed the integration with itself.

The scope of this change is only to move the code into the desired directory structure.
