+++ packages/@aws-cdk/aws-stepfunctions-tasks/lib/stepfunctions/start-execution.ts
@@ -1,7 +1,7 @@
 import * as iam from '@aws-cdk/aws-iam';
 import * as sfn from '@aws-cdk/aws-stepfunctions';
 import { Stack } from '@aws-cdk/core';
-import { getResourceArn } from './resource-arn-suffix';
+import { getResourceArn } from '.././resource-arn-suffix';
 
 /**
  * Properties for StartExecution
